requirements.txt
# 🔧 FastAPI Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0

# 🗄️ Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# 📊 Financial Analysis
pandas>=2.0.0  # More flexible version
numpy>=1.24.0
scipy>=1.11.0
yfinance>=0.2.28

# 🧪 Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2